+++ src/app/(pages)/courses/page.jsx
@@ -1,53 +1,8 @@
-'use client';
-
-import React, { useState } from 'react';
-import CourseCard from './CourseCard';
-import styles from './page.module.css';
-import courseData from './courseData';
+import React from 'react';
 
 function Courses() {
-  const [filter, setFilter] = useState('all');
-
-  const filteredData = courseData.filter((course) => {
-    if (filter === 'completed') {
-      return course.progress === 5;
-    } if (filter === 'incomplete') {
-      return course.progress < 5;
-    }
-    return true;
-  });
-
   return (
-    <>
-      <h1>Courses</h1>
-      <div className={styles.courseContainer}>
-        <div className={styles.courseToggle}>
-          <div
-            className={filter === 'all' ? styles.active : ''}
-            onClick={() => setFilter('all')}
-          >
-            All Courses
-          </div>
-          <div
-            className={filter === 'completed' ? styles.active : ''}
-            onClick={() => setFilter('completed')}
-          >
-            Completed
-          </div>
-          <div
-            className={filter === 'incomplete' ? styles.active : ''}
-            onClick={() => setFilter('incomplete')}
-          >
-            Incomplete
-          </div>
-        </div>
-        <div className={styles.courseLayout}>
-          {filteredData.map((course) => (
-            <CourseCard name={course.name} duration={course.duration} progress={course.progress} />
-          ))}
-        </div>
-      </div>
-    </>
+    <h1>Courses</h1>
   );
 }
