--- src/app/(pages)/courses/page.module.css
@@ -1,55 +1,0 @@
-/* course layout */
-.courseLayout{
-    display: flex;
-    gap: 20px;
-}
-
-/* course toggle */
-.courseToggle {
-    display: flex;
-    gap: 20px;
-    color: gray;
-    border-bottom: 2px solid gray;
-    padding: 5px;
-    margin-bottom: 30px;
-}
-
-.active {
-    font-weight: 900;
-}
-
-
-/* course cards */
-.courseCard{
-    display: flex; 
-    flex-direction: column;
-    justify-content: space-between;
-    border: 2px solid black;
-    border-radius: 10px;
-    padding: 20px;
-    width: 210px;
-    height: 140px;
-}
-
-.courseName{
-    font-weight: 500; 
-    font-size: 20px;
-    width: 100px;
-    border-left: 4px solid;
-    padding: 0px 10px;
-}
-
-.bottomBar{
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-}
-
-.courseDuration{
-    border-radius: 10px;
-    width: 40px;
-    background-color: rgb(207, 187, 187);
-    text-align: center;
-    font-size: 14px;
-    padding: 2px 5px;
-}
