+++ src/app/(pages)/landing/CourseCard.js
